+++ src/main/resources/templates/auth/sign-up.html
@@ -13,7 +13,7 @@
         <div class="col-6">
           <h4 class="mb-3 text-center">회원가입</h4>
           <hr>
-          <form id="userSignUpForm" action="/user/sign-up" method="post">
+          <form id="userSignUpForm" action="/api/user/sign-up" method="post">
             <div class="row g-3">
               <div class="mb-3">
                 <div class="input-group">
